fix pylint
